Change item rating datatype

+++ backEnd/models/itemModel.js
@@ -33,7 +33,7 @@ Item.init(
     },
 
     rating: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(2, 1),
       allowNull: false,
       defaultValue: 0,
     },
